fix(app): resolve DropdownMenu ColumnScope and SkipIntroLengthDialog type mismatch

+++ app/src/main/java/eu/kanade/tachiyomi/ui/entries/anime/AnimeScreen.kt
@@ -375,7 +375,7 @@ class AnimeScreen(
                     initialSkipIntroLength = if (!successState.anime.skipIntroDisable &&
                         successState.anime.skipIntroLength == 0
                     ) {
-                        screenModel.gesturePreferences.defaultIntroLength().get()
+                        screenModel.gesturePreferences.defaultIntroLength().get().toLong()
                     } else {
                         successState.anime.skipIntroLength
                     },
